#29 - Optimising blog wall title rebuilds
